Implement search feature for mobile
