+++ src/components/Button/Button.module.css
@@ -21,7 +21,6 @@
     transition: all 180ms ease-in-out;
 
     &:hover {
-        opacity: 0.7;
         transform: scale(1.01);
     }
