+++ src/components/Header/Header.jsx
@@ -1,10 +1,11 @@
-import { IoSearch as SearchIcon, IoCartOutline as CartIcon, IoMenu as MenuIcon, IoStorefrontOutline as ShopIcon } from "react-icons/io5";
-import Button from '../Button/Button';
+import { IoCartOutline as CartIcon, IoStorefrontOutline as ShopIcon } from "react-icons/io5";
 import Logo from "../Logo/Logo";
 import styles from './Header.module.css'
-import { Link, useNavigate } from "react-router";
 import { useCartContext } from '../../context/CartContext';
 import { useEffect, useRef} from "react";
+import { Link } from "react-router";
+import SearchBar from "../SearchBar/SearchBar";
+
 
 
 function Header() {
@@ -13,36 +14,12 @@ function Header() {
     return (
         <header className={styles.header}>
             <Link to="/"><Logo /></Link>
-            <SearchBar /> 
+            <div className={styles.searchBarContainer}>
+                <SearchBar />
+            </div>
             <Nav />
         </header>
     );
-}
-
-export function SearchBar() {
-    const navigate = useNavigate();
-
-    function handleSubmit(e) {
-        e.preventDefault()
-        const searchTerm = e.target.search.value;
-        e.target.search.value = '';
-        navigate(`/store?search=${encodeURIComponent(searchTerm)}`)
-    }
-
-    return (
-        <form className={styles.searchBar} onSubmit={handleSubmit}>
-            <input 
-                type="text" 
-                name="search"
-                placeholder="Search products..."
-                autoComplete="off"
-                className={styles.input}
-            />
-            <button type="submit" className={styles.inputBtn} >
-                <SearchIcon />
-            </button>
-        </form>
-    )
 }
 
 export function Nav() {
@@ -87,6 +64,7 @@ export function Nav() {
         </nav>
     )
 }
+
 
 
 export default Header;
