+++ src/pages/Store/Store.module.css
@@ -7,11 +7,22 @@
 }
 
 .header {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 1rem;
     text-align: center;
 
     & h1 {
         font-size: 3rem;
     }
+}
+
+.searchBarContainer {
+    display: none;
+    width: 100%;
+    justify-content: center;
+    padding: 1rem;
 }
 
 .storeMain {
@@ -111,6 +122,10 @@
 
     .header h1 {
         font-size: 2rem;
+    }
+
+    .searchBarContainer {
+        display: flex;
     }
 
     .storeMain {
